IOS-4828 Moved LayoutAlignment extension
